feat(cli): monitors/dashboards-as-code + in-process CLI lifecycle tests
- Move cli/ into the library (CLI.Main.cliMain); exe is a thin Paths wrapper.
  Enables runCLILifecycle: real optparse parse + command pipeline in-process,
  HTTP routed to handlers via runHTTPtoServant, stdout captured (silently).
- monitors apply was not an upsert: an id-less file created a new row on
  every apply. Add POST /api/v1/monitors/apply (natural key: alert title,
  mirroring dashboards' file_path), wire CLI applyOne to it.
- Add monitors yaml ID dump (inverse of monitorFromInput) so dumps round-trip
  through apply; same contract dashboards already had.
- Test router: thread verb + query params through writes (PUT/PATCH/POST were
  indistinguishable, mute --for lost its duration); add monitors, dashboards,
  events, facets, metrics, share routes + DELETE routing.
- CLILifecycleSpec: monitors-as-code and dashboards-as-code lifecycle walks
  (apply idempotency, yaml round-trip, mute/unmute/star/duplicate/delete).
